some bugs are worked out, but testing has hit a wall
the gtest unit tests don't currently support testing both cases for *_mem_map()
that is, an address can be passed in by the user (current behavior), or a NULL
can be passed in and the memory allocated by *_mem_map() and given back to the
user.  right now, only the NULL case is supported by the sysv transport, but
this case is not tested.