SM/MM: fix possible sending in out-of-order.
each sender would keep a local cached value for the value of
the remote peer's tail.
this value would be updated for pending sends to make sure that pending
operations are always scheduled before 'regular' sends (not from pending queue).
thus preventing out-of-order.
the cached value is also updated when in a 'regular' send and the
pending queue is empty.