UCT/SM/MM: allow the user to specify the path to the backing file when using posix.
the user can choose whether or not to use shm_open, when using posix, with
the following env parameter - UCX_POSIX_USE_SHM_OPEN (default is 'yes').
if shm_open was not chosen (or if it failed), open() will be used
instead. in that case the path to the backing file, that open() would use,
can be specified with UCX_POSIX_DIR (default is '/tmp').

Set default to 'try'.