CONFIG: Disable optimization by default
UCX can be compiled on machines with new cpu features available (avx, ...)
In this case it uses special optimization and as a result the
final binary is not able to be used/launched on machines unsupported
such features. Actually the behaivour could be unpredictable.

Make optimized binary explicitly using --with-avx, --with-sse41