UCP/GTEST: support Rndv for all contig/generic combinations.
+ added support for Rendezvous for all combinations of contig and
generic datatypes. i.e. the sender and receiver may have any of these
datatypes and rndv will be triggered starting the rndv threshold.

+ this also adds support for rendezvous with active messaging.
for example, for a case where both datatypes, on send and recv are generic,
for transports that cannot perform a get operation and when the recv
side a generic datatype.

+ test this in gtest.