UCS: Added an env param to set a log level for triggering error handling
The new env parameter - UCX_LOG_LEVEL_TRIGGER allows the user to set the
log level up to which he would like to trigger the error handling flow.
Up until now, it was the 'fatal' log level, now it's configurable.
Note: the UCX_LOG_LEVEL env param needs to be set to a value that would
allow the output of the messages that would be printed with the log level
set by the new parameter.