UCP/API: Add ucp_request_free() function - release and disable callback.
 In order to solve #1106, need to make sure that after a non-blocking
request is reset by the user, it will not be changed by the completion
callback. Solving this in the MPI library would require extra branches
to check request state, while we already do those checks in UCP layer.

 This commit adds new routine ucp_request_free() which works same way as
ucp_request_release() but also makes sure the completion callback for
the request will not be called. Users are encouraged to switch from
ucp_request_release() to ucp_request_free().

As part tof this commit following changes are made as well:
+ Add request flag which indicates if a callback should be called upon
  completion. Use it to disable callbacks for "external" requests and
  free'd requests.
+ Refactor request completion routines to make better use of new flags.
+ Add test for ucp_request_free().