UCT/DC: Fix hang in flush.
Currently, if a DC endpoint is waiting for grant reply, it is not
flushed until that grant arrives. This results in an endless loop when
trying to add the flush operation to pending queue (because pending_add
always returns BUSY). Even if the pending operations would have been
added, it would never progress because the arbiter makes progress when
getting a grant only if the window is 0 (to avoid reordering).
 As a result, we must allow destroying an endpoint with pending grant
request. However, we don't release its memory, so when the grant arrives
we would be able to find that endpoint.