UCS/RCACHE: Fix rcache merge while mappings are modified.
When merging with existing memory regions, it may already be released by
another thread, or even triggered by releasing some on rcache internal
structures. It's not always possible to detect this release on time
before attempting to register the region. Therefore, let the memory
registration callback report the error, and retry the registration with
the original address range (which has to be valid).