build: split ucx.pc into separate .pc files
A program wishing to some or all of UCX can easily and correctly
specify all its dependencies by

PKG_CHECK_MODULES([ucx], [ucs uct ucp])

or whatever other set of the uc* libraries it needs. This way, a
downstream user is explicit about what exact components it needs and
will not have to rely on a variadic list (which ucx.pc is) of
libraries which either contains too many libraries, or too few.

Fixes #1599.