UCP: On-demand progress wih signaled active messages.
If a transport interface supports only signaled active message events,
still use it for on-demand progress. The sender side will create a
signaling proxy endpoint, which will send the first active message as
signaled and then replace itself with the real endpoint.
In order to avoid the extra overhead in case active messages are not
used, RMA and AMO lanes could still contain a direct pointer to the
endpoint, even if it's also wrapped by a proxy for active messages.