UCT/GTEST: rdmacm md: check that can create an event_channel.
create a dummy event_channel during init so that we would know if rdmacm
can be used as an md in the future. a possible outcome of this check is
that the rdma_cm moudle isn't loaded and so the /dev/infiniband/rdma_cm
file, for example, cannot be access (and therefore rdmacm cannot be used).
need to detect this before selecting the rdmacm md for usage.