API/UCT: Sockaddr calls - server callback and uct_ep_create_sockaddr.
+ Remove the uct_sockaddr_conn_ready_callback_t callback from the server
side, since it's not always possible to have an indication from the
client side that the connection was established.
In case of UDP port space in rdmacm, for example, the client won't
send the server a message upon finalizing the connection established.

+ Change the return status of uct_ep_create_sockaddr.
Return UCS_INPROGRESS instead of UCS_OK since there is a callback on the
client side to indicate that the connection was established.