UCX: Fixing error handling for sysconf call
sysconf return value -1 does not necessary mean an error. For max/min
values it may mean that the value is undefined. According sysconf man
pages the right way to check for an error is to verify that errno value
was not changed. Also, we should not convert the value to unsigned, but
let the upper layer to handle undefined values for min/max variables.

On Ubuntu 17.10 _SC_IOV_MAX value is undefined and this leads
to a failure in gtest.

Signed-off-by: Pavel Shamis (Pasha) <pasharesearch@gmail.com>