UCP: Add 'aux' mark for auxiliary tls in alias
Some transport aliases in UCP contain additional tls intended to be
used for wireup process (for instance "rc" -> "rc, ud", where ud is
supposed to be used for wireup).
This commit adds special "aux" mark for such tls, which should guarantee
that this tls may be selected for wirep purposes only.