TOOLS/PERF: Fix a bug with batch file processing
Every time a batch file line is processes, need to revert to parent test
parameters, otherwise parameters from one test would be used as the
default parameters for the next test. It caused a hang in the test
because stream test which came after tag used iov datatype instead of
contig, miscalculated the expected receive length, and entered the RTE
barrier too early before the sender had a change to complete.