UCP/WIREUP: Fix hang due to wireup request message not being sent
If there are already pending operations on an endpoint, a wireup message
which is sent in order to obtain dest_ep_ptr may never be sent. This is
because the tranport executes the pending operations in order, and now
the queued requests are not progressing anymore because ep is switched
to wireup proxy mode. This prevents the transport from progressing the
wireup request, which results in a deadlock.
The fix is to extract those pending operations and requeue them after
the wireup request message.