UCS/UCT/MM: Prevent potential out-of-order sending in shared memory.
+ In case an endpoint fails to get ownership of the remote head element,
do not add the request to the pending queue (don't return a status that
would make the user add to pending). Instead, keep trying to get
ownership.
The failure to get ownership could be due to failing to achieve the
wanted outcome from the atomic operation, and NOT because the remote
peer doesn't have room in its fifo.
If it is added to the pending queue, the following sending ep may
succeed to get ownership and its message would be sent before the one
in the pending queue, thus, creating out-of-order sending.

+ ppc arch - update the macro for ucs_memory_cpu_load_fence()

(cherry picked from commit 17cfda619e9a822e9120845333aec9443e561fa6)