UCS/ARCH: Read from `cpu MHz` for CPU speed on x86
When the user changes the frequency of the CPU, UCS
still reads the base frequency on x86 for its
calculations.

This patch ensures that the CPU frequency is read
from the `cpu MHz` field of `cpuinfo` and not from
`model name`. The function that reads from the
model name is removed because it is not used anywhere
else and, if kept, would throw a build error.

Signed-off-by: Rohit Zambre <rzambre@uci.edu>