UCS/UCM: Avoid linking UCM loadable modules with "nodelete" flag
When a shared library is linked with "-Xlinker -z -Xlinker nodelete"
flags, it's not cleaned up properly when it fails to load. For example,
libucm_cuda.so can fail to load if CUDA is not present during runtime.
As a result, the library is still listed in process' link map, but it's
l_versions field is NULL. This leads to a segfault in glibc dlopen()
code, which in some cases is invoked from pthread_cancel_init().

So, instead of using linker flags, we use dlopen(RTLD_NODELETE). This
does not have such issue.