UCM: Avoid releasing environment strings in library destructor
Since now libucm.so is not linked with "nodelete", it may be unloaded
before it installs malloc hooks. In this case we would not have any
environment strings, so no need to release. If it did install hooks,
it means the library was aready reopened with RTLD_NODELETE, so no need
the strings anyway.

This fixes a case where libucm.so was unloaded before hooks were
installed, and called clearenv() which removed all environment variables
from the program, leading to segfault.

Since we are not clearing the environment now, we also cannot release
the allocated strings. This is not reported as memory leak by Valgrind,
since they are still reachable from the global array of environ strings.