UCT/API: A new client-server API in UCT with a connection manager.
The new API manages the client-server flow in UCT through the new entity
of a connection manager (cm).
A transport layer listener and endpoint (for a sockaddr connection)
will be opened in the context of a cm, rather than on an iface
which depends on an md.
The new API also adds new callbacks for handling connection
establishment on both client and server sides.
This is done as preparation for the addition of a close protocol to UCX.