UCP/API/TEST: Fix ucs_status_ptr_t checks
We can use UCS_PTR_STATUS() only if UCS_PTR_IS_PTR() is false or
UCS_PTR_IS_ERR() is true. Otherwise, must compare to NULL. This is
because converting a valid pointer to an enum may yield 0, if the
pointer is aligned to the size of the enum type.