UCS/ASYNC/TEST: Fix sync handler remove
In order to support removing async handler while its callback is called,
we must make sure the call happens from the same thread. Otherwise, the
sync-remove operation does not really guarantee sync, and subsequent
file operations (such as accept() in sockcm) will fail because fd is
already closed.

Refactor polling on missed events to avoid deadlock - we can assume the
polled handlers come from same async context we are polling on (because
of the way the are added to miss queue).

Add a test to make sure the async handler is not called after it's
sync-removed.
(cherry picked from commit 957c9de56171cdcf490396c6edda5330b6300866)