UCM/RELOC: Take original free() pointer from main program if possible
In some cases the previous value from libucm.so may not be correct
because of points to free@plt of the main program. This leads to
infinite recursion because we patched the main program to call
ucm_free() instead.
The fix is to take original free pointer from main program before we
patch it to call ucm_free(). Only if it doesn't exist we use libucm.so.
In addition, we need to ignore .plt reference when obtaining the
previous symbol value, since it does not point to the real function.