UCT/GTEST: tcp_sockcm ucp test fixes.
1. for the ucp err_handle sockaddr test, UCS_ERR_NOT_CONNECTED is a
valid return status, due to ECONNREFUSED in case the client tries to
connect to a non-existing port on the server side.
(added recently in the reject_reason change)

2. no need to print an error message in case of the EPIPE errno since
it's expected in case the remote peer closes its ep with the Force mode.

3. rename the GOT_DISCONNECT flag to UCT_TCP_SOCKCM_EP_DISCONNECTED, and
set it right before the upper layer's disconnect_cb is called.
thus, allowing the uct_tcp_sockcm_ep_disconnect function to continue
even if the Failed flag is set.