UCT/RC: Fix QP cleanup with tx_moderation enabled
Reserve a CQ entry for NOP operation during RC QP flush. Only one is
needed since this is a blocking operation. This removes the need to
disable TX moderation to avoid deadlock during flush due to lack of CQ
resources to post a single NOP (which is needed to gen completions for
unsignaled WQEs)