TEST/IODEMO: Fix increased memory usage due to invalid outstanding count
The client counts incorrectly the number of outstanding ops: when a
connection is closed, _num_sent is decreased, but it should not be,
since read requests also complete with error status and _num_completions
is increased. As a result, the client sends more and more outstanding
operations after every terminated connection, which causes increase in
memory usage.
The fix is to first close the connection, which will restore completions
for read requests, and then restore the remaining credits for write
requests - which will never complete on their own because they are
waiting for write completion io-message.