UCM: Fix mmap events test
Don't expand 'events' to native events when testing the installed mmap
events. Some events could be reported as "external", for example
VM_UNMAPED, in which case the corresponding native events are not
reported for it. We should test only the events which were passed
explicitly to ucm_test_events() API.