UCT/BASE: Fix deadlock in TCP
The macro uct_pending_queue_dispatch() did not handle ERR_NO_RESOURCE
correctly: it re-added the pending element without exiting the dispatch
loop. This leads to a deadlock in TCP (which uses this macro) since
there are no send resources, and progress is not being called to release
more send resources (because stuck in the pending loop).