UCP/API: Add request flag to influence latency vs. bandwidth protocol
Add request hint flag UCP_OP_ATTR_FLAG_MULTI_SEND which advises UCP that
that multiple requests are being sent in parallel, so prefer to use a
protocol which optimizes the overall parallel bandwidth.
Without this flag, the default selection will be done, which optimizes
the latency of passing the data to the remote side.