UCP/KEEPALIVE: Fix assertion that keepalive done too soon on same ep
Due to 0d591cc85a, if keepalive is currently in progress we restarted
the current round, which caused assetion failure that keepalive was done
on same ep (this case, first ep in the list) too often.

So if a keeaplive round is inprogress when we removed the last ep, we
should complete it. Otherwise, don't do anything.