UCS/SYS: Fix error logging for ENOSPC from shmget(2).
If `shmget(2)` fails with `ENOSPC`, it's supposed to print an error
message describing which limit was exceeded using a call to `shmctl(2)`.
At the moment, however, it confuses success and failure from that call
and so just immediately returns without generating the explanatory
message. This patch inverts the error checking for this call.