TEST/PTR_MAP: Coverity fix and refactor
The coverity fox for rdma_cm caused a coverity error in ptr_map: lock
order inversion when using UCS_CLEANUP_ONCE.
We solve it by making the ptr_map and mutex class member variables,
instead of global variables. So they could be initialized in the
ctor/dtor functions of the test case.