UCT/UD: Fix UCT completion not called leak during uct_ep_destroy()
Before this fix, the iface async completion queue dispatch could remove
a completion desc from the queue but not call the upper layer callback.
It can happen when we destroy two ud endpoints, and the queue contains
completions for both of them: we would remove all completions, but call
the callback only for the first destroyed endpoint.

As a result, some UCP requests were not released wnd lanes are
destroyed.

The fix is to remove only the relevant completion descs from the queue.