UCP/API/INFO: Add API to define endpoint scope name for debug info
Endpoint scope name improves transports and protocols debug info, by
assigning a name to endpoint configuration. For example, can see which
transports are used for inter-node and which transports are used for
intra-node.

In addition, this commit improves indentation when printing endpoint
information, and hides internal endpoints (such as memtype).