UCT/ROCM/IPC: use remote_agent if available
use the correct remote_agent handle if information is available. Using the local_agent for both src and destination is just the backup plan in case the remote_agent info is not available, e.g. in case the user limited the visibility of devices for a/some/all processes.

(cherry picked from commit 79ea940bebe9308366dd467567f6867ccdcd16b7)