UCP/CONTEXT/CONFIG: Use ucs_config_parser_clone_opts for copying config
Instead of using direct-assignment to initialize context->config.ext and
additionally duplicate configuration strings, use the configuration parser
deep-copy method ucs_config_parser_clone_opts().

In order to use it, needed to separate the context configuration table
to the "outer" table that contains all configuration fields, and the
inner "ucp_context_config_table" that contains just the fields we want
to save on the context instance.