UCP/WIREUP/SELECT: Update wireup lane bandwidth calculation
Add weighted sum of the local and remote iface bandwidths.
Using only the minimum of local and remote bandwidths can result in
the same score for all lanes if the local (remote) ifaces have a lower
bandwidth than all remote (local) ifaces. This can lead to undesirable
lane selections because the NIC-device affinity on the remote (local)
side is ignored.