ARCH: Adding support for Data Gathering Hint
This patch aligns UCX with kernel subsystem:
https://patchwork.kernel.org/project/linux-rdma/patch/0-v1-c5dade92f363+11-mlx5_io_stop_wc_jgg@nvidia.com/

The patch streamlines the code quite a bit. Since DGH is a hint
instruction, on platforms that do not support the DGH extension the hint gets
translated to no-op.

It is safe to remove TX2 branch since anyways it is translated to  no-op.

It is also safe to remove DMB barrier because on older implementations the
buffer gets retired before 64B write is completed. Alternatively, it can
be retired on timeout and on most platforms it will be faster than DMB.

Signed-off-by: Pavel Shamis (Pasha) <pasharesearch@gmail.com>