BUILD/DEBIAN: Use dep libnvidia-ml1 for ucx-cuda
The names of packages provided by NVidia in CUDA don't seem to follow
the standard required by Debian (package per library), and therefore
shlibdep creates dependencies that are too strict: on
libnvidia-compute-515 instead of the actual name of the library.

Thus this commit fixes the generated dependency: adding "|
libnvidia-ml1" (that is: may also be satisfied by any package that
provides libnvidia-ml1).

It seems that the name of the library, provided by all relevant
packages, is "libnvidia-ml1". Use the standard dependency name of
<specific> | <provides>.