GTEST: Limit max gtest message size to BAR1 size
Current max message size does not take in consideration the limitations
of BAR1 size. For ML GPUs the BAR1 is very large it has never been a
limitation. For consumer class of GPUs the BAR1 size can be in a range
of megabytes. As a result gtest used to fail. This patch add query for
the BAR1 size and limits the max message size.