UCP/EP: Fix rendezvous threshold for multi-path
When using multiple paths on the same resource, the bandwidth of each
path should be estimated as a fraction of the total bandwidth.

Without this fix the bandwidth is assumed to be more than it really is,
which sets the rendezvous threshold to a lower value than needed.

This fix applies only to protocols v1.