TEST/UCP: Separate worker progress to avoid concurrent `ucp_worker_progress()`
Without `--enable-mt`, there are asserts on tests like `test_ucp_fence*`:
```
==>  2807     ucs_assert(worker->inprogress++ == 0);
```

The sender calls `request_wait()` or `flush_worker()` that eventually calls
`check_events()`. That later one progresses all entities, so the sender and the
receiver workers. This is an issue because the main thread is already
progressing the receiver worker.

To fix, we make sure that the sender thread does not progress the
receiver worker by adding one new parameter to restrict processing to a subset
of the list of entities. In case of `self`, only the sender needs to progress.