UCT/ROCM: revert a5246b8 for rocm_ipc
Commit a5246b8 moved the memhandle cache for rocm_copy and rocm_ipc from
the endpoint datastructure to the iface data structure. While this is
correct and benefitial for rocm_copy, it is actually detremental
for rocm_ipc.

The reason is that the memhandle cache is using virtual addresses for
finding and identifying a registered memory region and ipc key.
Multiple processes can use however the same virtual address, in which
case the memhandle cache will find entry, but will fail in the comparison
of the ipc keys stored. Consequently, the operation will unmap the previous
key, and map/open the current ipc key. Compared to the previous solution
this can lead to thrashing of ipc registrations in case the same
virtual address is used by multiple processes.

This commit therefor revert commit a5246b8 for rocm_ipc only, but keeps
the changes for rocm_copy.