UCM/UTIL/RELOC: Don't assume dynamic table values are pointers
On some combinations of C library implementation and architecture the
entries in the PT_DYNAMIC table are not relocated and stored as
pointers. Instead the values are offsets from the address that the
library is loaded at.

This was discovered when using UCX_MEM_MMAP_HOOK_MODE=reloc on RISC-V
however it is not limited to RISC-V only as some other C library
implementations (like musl) will also not update the dynamic table with
pointers

Unfortunately since the internal glibc headers that expose whether the
value is relocated on not are not available the best solution is to use
a heuristic to determine if the values needed from the table (strtab,
symtab, etc) are offsets or pointers is by comparing against the
library start address.

Since this should only be done for the values from the table that are
treated like pointers create a new function ucm_reloc_get_pointer() that
applies the heuristic.

Fixes: #9283

Signed-off-by: Evan Green <evan@rivosinc.com>
Co-developed-by: Rob Bradford <rbradford@rivosinc.com>
Signed-off-by: Rob Bradford <rbradford@rivosinc.com>