UCT/IB/MLX5: Don't create KSM with atomic access unless needed
On machines with relaxed-order PCIe enabled by default, we need to
create a strict order key when fence-imposed ordering is needed.
If atomic operations are not really used (for example, MPI use-case),
we don't really need to create a KSM with atomic access. We can just
use the strict-order direct key.
This avoids the extra overhead and potential failure of creating an
indirect key with atomic access.