GTEST: use stdc++17 for compiling hip tests
hip_runtime.h contains in some rocm releases a C++17 feature.
Prior to this commit, we forced using gnu++11 using HIP_CXXFLAGS,
but combined with the -Werror flag of UCX this leads to a compilation
error.

This commits adds the configure logic to check for the std=c++17 flag,
and uses this flag instead of gnu++11 if available. Forcing using the
std=c++17 flag revealed a second issue however regarding ignoring error
codes. This commit fixes this as well.