UCM/BISTRO: Use constant size in memcpy for ucm_bistro_modify_code
By using the constant value from sizeof() we can be sure that the
compiler will not generate an out of bounds error even for code that is
unreachable.

e.g If sizeof(ucm_bistro_lock_t) == 4 the compiler might generate an out
of bounds error for the copy of bytes into the uint16_t (as 4 > 2) even
though that code is unreachable.

Fixes: #9269

Signed-off-by: Rob Bradford <rbradford@rivosinc.com>