UCT/ROCM: don't try to lock hipHostMalloc'd memory
with v2 protocols there seems to be a scenario where we try to register
hipHostMalloc'd memory and fail.  This fix here ensure that we
skip the hsa_amd_memory_lock() operation during memory registration
for all memory types that return HSA_EXT_POINTER_TYPE_HSA (which
includes device and hipHostMalloc'd memory).

With this fix we are down to one single test that does not pass with v2
protocols :-) (which passes with v1).