UCT/ROCM: don't check buffer size during key-packing
The ROCr runtime returns in the hsa ptr_info function as the buffer size the exact
length that was allocated. The length provided as an input to the rocm_ipc_pack_key
routine might however have been adjusted by up to 64 bytes (if my math is correct) when
using the rcache. So comparing the input buffer lengths vs. the size
returned by the hsa_ptr_info function can be inaccurate. In addition, I do
not see what the comparison of the 'expected' size vs. 'real' size
gives us from the functionality perspective in this routine.