UCT/ROCM: set parameters dependent on GPU type
set the bandwidth values in the rocm/copy performance estimate
functions based on the GPU type that we found. We use the same
parameters for MI300A and MI300X at the moment, but the
infrastructure is in place in case we need to distinguish between
them at a later point.