UCS/CPU: Make headers C89 compliant
This is not a correctness issue but a user friendliness issue. Including some UCX headers while compiling OMPI with picky compiler options generate a ton of warnings.

The warning in compiler_def.h is about the two branches of the conditional returning different integer types (one int and the other unsigned long).

The other warnings are about declaration of functions with no arguments. The solution here is either adding void to signal there are no arguments or making the declaration a prototype.

Fixes #9910

Signed-off-by: George Bosilca <gbosilca@nvidia.com>