UCT/IB/UD: Use GRH to detect address family on non-Mellanox hardware
Setting the service level in the work completion is a Mellanox-specific feature,
so it can not be relied on to detect IPv4 vs IPv6. This commit fixes the
detection logic for non-Mellanox providers by detecting the address class from
the grh instead. This is done by detecting either 0x6a (IPv6) at offset 0 or
0x45 (IPv4) at offset 20 of the receive buffer. Since the first 20B of IPv4
packets are undefined ud_verbs sets the first byte of each posted receive to a
known value (0xff) since the provider is unliklely to touch these bytes. This
commit makes no changes to the mlx5 code which continues to rely on the CQE data
to determine if a packet is IPv4 or IPv6. It can be updated to use the non-mlx5
logic but since the IP version is present in the CGE there is no need.

Signed-off-by: Nathan Hjelm <hjelmn@google.com>