UCP/TEST/TAG: Fix the limit on number of endpoint lanes
To make the tag offload multi test pass, we must limit the number of
lanes on an endpoint to 1. Just limiting the number of eager and
rendezvous lanes is not enough, since different lanes could be selected
for eager and rendezvous protocols.
We add a new configuration parameter to also limit the total lanes on
and endpoint.