CONFIG: remove bad macro that overwrote automake internals
In the original port to autotools, a "ts" macro was defined in m4. This
meant any whole word "ts" in configure.ac or any included *.m4 file got
replaced with an integer plus a newline char. Automake does this
sometimes, such as in:

```
rm -f conftest.ts?
```

which was translated to:

```
rm -f conftest.20250527230813
?
```

At the time it was used to replace, on a later line,
```
BUILD_TS=ts
```

But this was removed as "unused" in 2017. It was always the wrong way to
define it and is now inert except for breaking automake. Get rid of it
fully.

Fixes: 597216e41e0b86219f87b965b18b1622247006c4
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>