UCT/ROCM: force flush to finish all pending ops
This PR fixes an issue that was uncovered while testing UCX 1.19.0-rc1.
The symptom is that the osu_get_bw tests fail with device memory.
Specifically, UCX does not seem 'wait' during the
window completions for the pending operations, and the queue of pending
ipc transfers is getting longer and longer until we either run out of
ipc signals or we segfault because the osu benchmark is releasing the
buffer while data transfers are still ongoing.

Making uct_rocm_ipc_iface_flush force to complete all pending operations
(instead of potentially just indicating with UCS_INPROGRESS that there
are more pending operation) seems to resolve the issue, though I suspect
that there are also other solutions.

This will also need to go to the 1.19 branch.